Imported fonts and styling for box icons.
