+++ client/src/app/layout.tsx
@@ -1,22 +1,13 @@
 import type { Metadata } from 'next';
-import { Fira_Sans, Sofia_Sans_Condensed } from 'next/font/google';
+import { firaSans, sofiaSansCondensed, boxIcons } from './fonts/fonts';
 import ThemeProvider from '../context/themeProvider';
 import NavBar from './components/navBar';
 import './styles/globals.scss';
-
-const firaSans = Fira_Sans({
-  weight: ['400', '500', '700'],
-  subsets: ['latin'],
-  variable: '--fira',
-});
-const sofiaSansCondensed = Sofia_Sans_Condensed({
-  subsets: ['latin'],
-  variable: '--sofia',
-});
+import './styles/boxicons.css';
 
 export const metadata: Metadata = {
   title: 'MapHub',
-  description: 'Your one stop map editor ',
+  description: 'A complete map visuals studio.',
 };
 
 export default function RootLayout({
@@ -27,7 +18,7 @@ export default function RootLayout({
   return (
     <html 
       lang="en"
-      className={`${firaSans.variable} ${sofiaSansCondensed.variable}`}
+      className={`${firaSans.variable} ${sofiaSansCondensed.variable} ${boxIcons.variable}`}
     >
       <body>
         <ThemeProvider>
